+++ CHANGELOG.md
@@ -6,6 +6,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.3] - 2023-09-10
+
+### Fixed
+
+- Switched ytdl-core to patched version
+
+## [2.4.2] - 2023-08-12
+
+### Fixed
+- Bumped node-ytsr ([#948](https://github.com/codetheweb/muse/issues/948))
+
+## [2.4.1] - 2023-07-23
+
+### Fixed
+- Autocomplete suggestion search for `favorites use` command is no longer case-sensitive
+- Autocomplete suggestion results for `favorites use` could return >25 results which Discord's API does not support
+
+## [2.4.0] - 2023-07-19
+### Added
+- Pagination to the output of the `favorites list` command
+
+### Fixed
+- Favorites list exceeding Discord's size limit could not be
+  viewed ([#606](https://github.com/codetheweb/muse/issues/606))
+
+## [2.3.1] - 2023-07-18
+### Fixed
+- Bumped ytdl-core
+
 ## [2.3.0] - 2023-05-13
 ### Added
 - Muse now normalizes playback volume across tracks. Thanks to @UniversalSuperBox for sponsoring this feature!
@@ -242,7 +271,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[unreleased]: https://github.com/codetheweb/muse/compare/v2.3.0...HEAD
+[unreleased]: https://github.com/codetheweb/muse/compare/v2.4.3...HEAD
+[2.4.3]: https://github.com/codetheweb/muse/compare/v2.4.2...v2.4.3
+[2.4.2]: https://github.com/codetheweb/muse/compare/v2.4.1...v2.4.2
+[2.4.1]: https://github.com/codetheweb/muse/compare/v2.4.0...v2.4.1
+[2.4.0]: https://github.com/codetheweb/muse/compare/v2.3.1...v2.4.0
+[2.3.1]: https://github.com/codetheweb/muse/compare/v2.3.0...v2.3.1
 [2.3.0]: https://github.com/codetheweb/muse/compare/v2.2.4...v2.3.0
 [2.2.4]: https://github.com/codetheweb/muse/compare/v2.2.3...v2.2.4
 [2.2.3]: https://github.com/codetheweb/muse/compare/v2.2.2...v2.2.3